fix sim1134.v DIV & MUL to work overflow and carry bits 1ike real 11/34
test with testdiv.tcl and testmul.tcl

also turbo mode should only work when in simulator mode

+++ ccode/z11util.cc
@@ -289,7 +289,7 @@ uint32_t randbits (int nbits)
 {
     static uint64_t seed = 0x123456789ABCDEF0ULL;
 
-    uint16_t randval = 0;
+    uint32_t randval = 0;
 
     while (-- nbits >= 0) {
